Improved listen submit body
