+++ app/src/main/java/org/listenbrainz/android/di/ListensRepositoryModule.kt
@@ -1,14 +1,14 @@
 package org.listenbrainz.android.di
 
-import dagger.hilt.InstallIn
-import dagger.hilt.android.components.ActivityRetainedComponent
 import dagger.Binds
 import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
 import org.listenbrainz.android.repository.ListensRepository
 import org.listenbrainz.android.repository.ListensRepositoryImpl
 
 @Module
-@InstallIn(ActivityRetainedComponent::class)
+@InstallIn(SingletonComponent::class)
 abstract class ListensRepositoryModule {
     @Binds
     abstract fun bindsListensRepository(repository: ListensRepositoryImpl?): ListensRepository?
