+++ app/src/main/java/org/listenbrainz/android/repository/ListensRepositoryImpl.kt
@@ -3,14 +3,19 @@ package org.listenbrainz.android.repository
 import android.content.pm.PackageManager
 import android.graphics.drawable.Drawable
 import androidx.annotation.WorkerThread
+import okhttp3.ResponseBody
 import org.listenbrainz.android.application.App
 import org.listenbrainz.android.model.CoverArt
 import org.listenbrainz.android.model.Listen
+import org.listenbrainz.android.model.ListenSubmitBody
 import org.listenbrainz.android.model.TokenValidation
 import org.listenbrainz.android.service.ListensService
+import org.listenbrainz.android.util.Log.d
 import org.listenbrainz.android.util.Resource
 import org.listenbrainz.android.util.Resource.Status.FAILED
 import org.listenbrainz.android.util.Resource.Status.SUCCESS
+import retrofit2.Call
+import retrofit2.Response
 import javax.inject.Inject
 import javax.inject.Singleton
 
@@ -77,5 +82,20 @@ class ListensRepositoryImpl @Inject constructor(val service: ListensService) : L
         } catch (e: PackageManager.NameNotFoundException) {
             packageName
         }
+    }
+    
+    
+    override fun submitListen(token: String, body: ListenSubmitBody) {
+        service.submitListen("Token $token", body)?.enqueue(object : retrofit2.Callback<ResponseBody?> {
+            override fun onResponse(call: Call<ResponseBody?>, response: Response<ResponseBody?>) {
+                d("Listen submitted successfully.")
+                d(response.message())
+                d(response.code().toString())
+                d(response.errorBody().toString())
+            }
+            override fun onFailure(call: Call<ResponseBody?>, t: Throwable) {
+                d("Something went wrong: ${t.message}")
+            }
+        })
     }
 }
